pyproject.toml
[project]
name = "coinpilot-ai"
version = "0.3.0"
description = "Bitcoin Autonomous Intelligence Agent - 自動化加密貨幣分析與出版系統"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "CoinPilot Team" }]
dependencies = [
    "github-copilot-sdk>=0.1.0",
    "requests>=2.31.0",
    "feedparser>=6.0.10",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    # 技術指標分析
    "pandas>=2.2.0",
    "pandas-ta>=0.3.14b",
    "numpy>=1.24.0",
    # 新聞內容解析
    "newspaper3k>=0.2.8",
    "lxml>=5.0.0",
    "lxml-html-clean>=0.1.0",
    # Web GUI
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    # BAIA Agent 新增依賴
    "yfinance>=0.2.40",
    "mplfinance>=0.12.10b0",
    "matplotlib>=3.8.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
coinpilot = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "W"]
